+++ src/main/java/de/jeff_media/drop2inventory/config/Config.java
@@ -49,10 +49,12 @@ public class Config {
     public static final String WORKS_IN_CREATIVE = "works-in-creative";
     public static final String DROP_WHEN_INV_FULL = "drop-to-ground-when-inv-full";
     public static final String SAVE_PLAYERDATA = "save-playerdata";
+    public static final String SUPERIOR_SKYBLOCK_ONLY_ON_OWN_ISLANDS = "superior-skyblock-collect-only-on-own-islands";
 
     public Config() {
         Main main = Main.getInstance();
         FileConfiguration conf = main.getConfig();
+        conf.addDefault(SUPERIOR_SKYBLOCK_ONLY_ON_OWN_ISLANDS, false);
         conf.addDefault(IM_USING_OUTDATED_PLUGINS, false);
         conf.addDefault(WORKS_IN_CREATIVE, false);
         conf.addDefault(SAVE_PLAYERDATA, true);